refactor: Alterado a rota de atualização do usuário
